requirements.txt
dashscope
python-dotenv

openai